move utility methods from TypeGraph into TypeUtils
